backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
python-socketio==5.5.0
python-engineio==4.3.0
SQLAlchemy==1.4.23
requests==2.26.0
python-dotenv==0.19.0
websockets==10.0
pytest==6.2.5
flask-login==0.5.0
flask-migrate==3.1.0
flask-cors==3.0.10
